requirements.txt
streamlit
openai
langchain
faiss-cpu
python-docx
pypdf
python-pptx
pandas

